% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-create_calendar.R
\name{create_calendar}
\alias{create_calendar}
\title{Create a Scheduling Calendar}
\usage{
create_calendar(
  title = "Work Schedule",
  schedule = c("weekdays", "42", "5623", "custom"),
  start = "2021-01-01",
  end = "2021-12-31",
  anchor = start,
  cycle = NULL
)
}
\arguments{
\item{title}{The title of the calendar}

\item{schedule}{The schedule to use}

\item{start}{The start date of the returned schedule; either a string in
"YYYY-MM-DD" format or a `Date` object}

\item{end}{The end date of the returned schedule; either a string in
"YYYY-MM-DD" format or a `Date` object}

\item{anchor}{The date from which to start ("anchor") schedule calculations.
This can be any valid date; no particular relationship to `start` or `end`
is needed. It must be either a string in "YYYY-MM-DD" format or a
`Date` object.}

\item{cycle}{A logical vector defining one scheduling cycle. If named, names
are passed to \code{\link[schedules:parse_weekday]{parse_weekday()}} for
standardization of weekday names.}
}
\value{
A list defining the `calendR` object
}
\description{
`create_calendar()` creates a calendar from a given `schedule`, `start`
date, `end` date, and (if needed) `anchor` date. Custom schedules can be used
by setting `schedule = "custom"` and providing the scheduling cycle via
`cycle`. The schedule can be saved to pdf with `.pdf = TRUE`; the file name
is a standardized version of the title.
}
