% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-magrittr.R
\name{magrittr-helpers}
\alias{magrittr-helpers}
\title{Magrittr Pipe Helpers}
\description{
The magrittr pipe is a common programming tool in \strong{R} that often makes
code much easier to read. In addition to the primary pipe operator, magrittr
offers several additional pipes, as well as functions to help perform common
infix-operations in a pipe-friendly manner.
}
\keyword{internal}
