% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-rotations_to_cycle.R
\name{rotations_to_cycle-helpers}
\alias{rotations_to_cycle-helpers}
\alias{expand_rotations}
\alias{validate_rotations}
\alias{split_rotations}
\alias{std_rotations}
\alias{remove_whitespace}
\alias{dash_to_space}
\alias{keep_numeric_space}
\alias{space_to_dash}
\title{\code{\link[schedules:rotations_to_cycle]{rotations_to_cycle}} Helpers}
\usage{
expand_rotations(rtns)

validate_rotations(rtns)

split_rotations(rtns)

std_rotations(rtns)

remove_whitespace(string)

dash_to_space(string)

keep_numeric_space(string)

space_to_dash(string)
}
\arguments{
\item{rtns}{The `rotations` vector in various formats}

\item{string}{A character string}
}
\description{
These functions assist `rotations_to_cycle()`:

\itemize{
  \item{`std_rotations()` standardizes the `rotations` string}
  \item{`split_rotations()` splits the string into a vector of numbers}
  \item{`validate_rotations()` validates expectations for `rotations`}
  \item{`expand_rotations()` expands the numeric vector to a logical one}
}

`std_rotations()` has helper functions of its own:

\itemize{
  \item{`remove_whitespace()` removes all whitespace from a string}
  \item{`dash_to_space()` converts all dashes to spaces}
  \item{`keep_numeric_space()` removes all non-numeric/space characters}
  \item{`space_to_dash()` converts all spaces back to dashes}
}
}
