% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-rotations_to_cycle.R
\name{rotations_to_cycle}
\alias{rotations_to_cycle}
\title{Convert Rotations Input to an Expanded Shift Cycle}
\usage{
rotations_to_cycle(rotations)
}
\arguments{
\item{rotations}{Character. A string of alternating numbers and dashes
representing the "on" and "off" periods of a rotating schedule. This
must contain an even number of numerals and dashes between each numeral
(i.e. "4-2-3-3" is a valid input; "4-2-3" or "4 2 3" is not).}
}
\description{
`rotations_to_cycle()` converts a string like `"4-2"` to a vector of `TRUE`
and `FALSE` values to pass to
\code{\link[schedules:schedule_by_cycle]{schedule_by_cycle()}}
}
