% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-std_dates.R
\name{std_dates}
\alias{std_dates}
\title{Parse Dates to Standard Format}
\usage{
std_dates(x, orders = c("dmy", "mdy", "ymd", "dmyT", "mdyT", "ymdT"))
}
\arguments{
\item{x}{A vector of dates in a `Date`- or `POSIXct`-like format, or as
character strings}

\item{orders}{The orders to use when parse character strings with
\code{\link[lubridate:parse_date_time]{parse_date_time()}}}
}
\value{
A `Date` or `POSIXct` vector
}
\description{
`std_dates` standardizes a date vector and returns a vector in `Date` or
`POSIXct` format, depending on whether there is sub-daily information
available in the data.
}
