% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-std_names.R
\name{std_names}
\alias{std_names}
\title{Parse Investigator Names and Remove (Most) Non-Alphabetical Characters}
\usage{
std_names(string)
}
\arguments{
\item{string}{A character vector containing investigator names}
}
\value{
The input `string` with cleaned names
}
\description{
`std_names()` parses investigator names given in the teams
worksheet into a standard format for matching with REDcap. It:
\enumerate{
  \item Replaces square brackets and curly braces with parentheses
  \item Removes parenthetic expressions
  \item Removes ordinal and numeric expressions
  \item Converts `NA`s to text (for the next function)
  \item Passes the intermediate to
    \code{\link[janitor:make_clean_names]{
      make_clean_names(
        case = "title",
        use_make_names = FALSE,
        transliterations = c("Any-Latin", "Latin-ASCII"),
        sep_out = " "
      )
    }}
  \item Removes leading and trailing dashes
  \item Removes extraneous whitespace
  \item Inverts converted `NA`s (back to `NA_character_`)
}
}
\seealso{
Other Case Assignment: 
\code{\link{schedule_predefined}()},
\code{\link{schedule}()}
}
\concept{Case Assignment}
