% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-std_names.R
\name{str_invert_na}
\alias{str_invert_na}
\title{Invert `NA` String Replacement}
\usage{
str_invert_na(string)
}
\arguments{
\item{string}{A character vector}
}
\value{
The input `string` with `NA` replacements converted back to
  `NA_character_`
}
\description{
`str_inv_replace_na()` performs the inverse operation of
\code{\link[stringr:str_replace_na]{str_replace_na()}}; that is, it converts
replaced `NA` values back to `NA_character_`. This only works if `NA` is
replaced with `"NA"` or a case-wise variant thereof (i.e. `"Na"`, `"na"`, or
`"nA"`). It is designed to invert the addition of "`_[0-9]`" suffixes as
well.
}
\keyword{internal}
