% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-std_names.R
\name{str_remove_numeric}
\alias{str_remove_numeric}
\title{Remove Numeric Symbols from String}
\usage{
str_remove_numeric(string)
}
\arguments{
\item{string}{A character vector}
}
\value{
The input `string` with numeric symbols removed
}
\description{
`str_remove_numeric()` removes numeric symbols from a string. It also removes
decimals and commas within numeric strings.
}
\keyword{internal}
