% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-std_names.R
\name{str_remove_ordinal}
\alias{str_remove_ordinal}
\title{Remove Mixed Ordinal Numbers with Text Representation}
\usage{
str_remove_ordinal(string)
}
\arguments{
\item{string}{A character vector}

\item{num_paste}{A logical. If `TRUE`, the elements of larger numbers are
separated with spaces. If `FALSE`, the elements will be joined without
spaces.}

\item{A}{logical. If `TRUE`, ordinal numbers are removed from `string`}

\item{...}{Ignored.}
}
\description{
Removes mixed text/numeric ordinal numbers (i.e. "1st", "2nd", "23rd"). This
function was inspired by the `replace_ordinal()` function in the
textclean package, and re-uses pieces of its code.
}
\details{
Currently only implemented for ordinal values 1 through 100.
}
\keyword{internal}
