% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-std_names.R
\name{str_remove_parenthetic}
\alias{str_remove_parenthetic}
\title{Remove Text Surrounded by Parentheses}
\usage{
str_remove_parenthetic(string)
}
\arguments{
\item{string}{A character vector}
}
\value{
The input `string`, with parenthetic expressions removed
}
\description{
`str_remove_parenthetic()` removes parenthetic expressions from strings
(i.e. "Hi, Alice (and Bob)!" becomes "Hi, Alice !").
}
\keyword{internal}
