% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-std_names.R
\name{str_replace_brackets}
\alias{str_replace_brackets}
\alias{str_replace_braces}
\title{Replace Square Brackets and Curly Braces with Parentheses}
\usage{
str_replace_brackets(string)

str_replace_braces(string)
}
\arguments{
\item{string}{A character vector}
}
\value{
The `string`, with square brackets or curly braces replaced by
  parentheses
}
\description{
`str_replace_brackets()` and `str_replace_braces()` are convenience functions
for replacing square brackets and curly braces with parentheses.
}
\keyword{internal}
