% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-std_names.R
\name{str_trim_dashes}
\alias{str_trim_dashes}
\title{Remove Leading and Trailing Dashes from String}
\usage{
str_trim_dashes(string)
}
\arguments{
\item{string}{A character vector}
}
\value{
The input `string`, with leading and trailing dashes removed
}
\description{
`str_trim_dashes()` removes leading and trailing dashes from a string. This
is most useful to clean up the results of previous operations that may have
left such dashes.
}
\keyword{internal}
