% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-anchor_date.R, R/ui-calendar_title.R,
%   R/ui-date_range.R, R/ui-download_button.R, R/ui-rotations.R
\name{ui-utils}
\alias{ui-utils}
\alias{anchor_date}
\alias{year_first_monday}
\alias{calendar_title}
\alias{date_range}
\alias{year_start}
\alias{year_end}
\alias{download_dropdown}
\alias{download_button}
\alias{rotations}
\title{UI Helper Functions}
\usage{
anchor_date(id)

year_first_monday(date = lubridate::today())

calendar_title(id)

date_range(id)

year_start(date = lubridate::today())

year_end(date = lubridate::today())

download_dropdown(id, inputId = "download_dropdown")

download_button(
  id,
  inputId = "download",
  label = toupper(gsub("_", " ", inputId))
)

rotations(id)
}
\arguments{
\item{id}{ID assigned to the module calling the function}

\item{date}{A date to use for determining the year; defaults to
\code{\link[lubridate:today]{today()}}}

\item{inputId}{ID to give to this instance of the function}

\item{label}{A user-facing label for the UI element}
}
\description{
These functions wrap or assist shiny UI functions.
}
