% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-weekday_switch.R
\name{weekday_switch}
\alias{weekday_switch}
\title{Create a Material Switch for a Day of the Week}
\usage{
weekday_switch(
  id,
  day = c("sun", "mon", "tue", "wed", "thu", "fri", "sat"),
  value = FALSE
)
}
\arguments{
\item{id}{The id of the Shiny module calling this function}

\item{day}{A day of the week, as a 3-letter abbreviation in title case}

\item{value}{TRUE or FALSE.}
}
\value{
A switch control that can be added to a UI definition.
}
\description{
`weekday_switch()` creates a switch using
\code{\link[shinyWidgets:materialSwitch]{materialSwitch()}} for a given day
of the week. It labels the switch on the right side with the full weekday
name and lowercases the 3-day abbreviation for the `inputId`. This
is designed to work inside Shiny modules.
}
