% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-weekday_switch.R, R/utils_weekly.R
\name{weekday_switch}
\alias{weekday_switch}
\title{Create a Material Switch for a Day of the Week}
\usage{
weekday_switch(
  day = c("Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"),
  value = FALSE
)

weekday_switch(
  day = c("Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"),
  value = FALSE
)
}
\arguments{
\item{day}{A day of the week, as a 3-letter abbreviation in title case}

\item{value}{The default button value, `TRUE` or `FALSE`}
}
\value{
A switch control that can be added to a UI definition.
}
\description{
`weekday_switch()` creates a switch using
\code{\link[shinyWidgets:materialSwitch]{materialSwitch()}} for a given day
of the week. It labels the switch on the right side with the full weekday
name and lowercases the 3-day abbreviation for the `inputId`.

`weekday_switch()` creates a switch using
\code{\link[shinyWidgets:materialSwitch]{materialSwitch()}} for a given day
of the week. It labels the switch on the right side with the full weekday
name and lowercases the 3-day abbreviation for the `inputId`.
}
